dom/html/test/test_bug677463.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=677463
-->
<head>
  <title>Test for Bug 677463</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=677463">Mozilla Bug 677463</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 677463 **/

var o = document.createElement("option");
var m = document.createElement("menuitem");
is(o.label, m.label, "Should have same labels");
o.textContent = "   ";
is(o.label, m.label, "Should have same labels");
m.textContent = " ";
is(o.label, m.label, "Should have same labels");
o.textContent = " foo";
isnot(o.label, m.label, "Shouldn't have same labels");
m.textContent = "foo ";
is(o.label, m.label, "Should have same labels");
m.label = "bar";
isnot(o.label, m.label, "Shouldn't have same labels");
o.label = "bar";
is(o.label, m.label, "Should have same labels");

</script>
</pre>
</body>
</html>
